% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICcond.R
\name{AICcond}
\alias{AICcond}
\title{Computation of the Conditional Akaike Information Criterion (AICcond) for a joint model estimated by JointMult function}
\usage{
AICcond(model, Y, D, data, var.time, RE = "block-diag", BM = "diag", B,
  posfix, breaks = NULL, delayed = TRUE)
}
\arguments{
\item{model}{a JointMult model}

\item{Y}{a list of \code{multlcmm} objects}

\item{D}{a list of two-sided formula defining the event part of the model}

\item{data}{data.frame containing the observations and variables}

\item{var.time}{a character vector indicating the name of the time variables}

\item{RE}{an indicator of the random effect structure between dimensions}

\item{BM}{an indicator of the correlation of the Brownian motions}

\item{B}{vector containing initial values for the parameters}

\item{posfix}{optional vector specifying the indices in vector B of the parameters that are not estimated}

\item{breaks}{optional vector specifying the break points in the case where the event time is discretized}

\item{delayed}{logical vector indicating if delayed entry should be accounted for}
}
\value{
A list containing : \item{AICcond}{the conditional Akaike Information Criterion} 
\item{vrais2i}{a vector containing individual contributions to the conditional and total log-likelihood}
\item{npm}{the number of estimated parameters for the joint model}
\item{npmtot}{a vector containing the number of estimated parameters of each longitudinal submodel}
}
\description{
Computation of the Conditional Akaike Information Criterion (AICcond) for a joint model estimated by JointMult function
}
\references{
Zhang et al, Assessing model fit in joint models of longitudinal and survival data 
with applications to cancer clinical trials. Statistics in medicine 2014 vol. 33, no 27, p. 4715-4733.
}
\author{
Tiphaine Saulnier, Cecile Proust-Lima and Viviane Philipps
}
