\name{multLMP}
\alias{multLPM-package}
\alias{multLPM}
\docType{package}
\title{
Joint model for multivariate latent processes and clinical endpoints using an exact likelihood inference method
}
\description{
The package estimates a joint model for multivariate latent processes underlying longitudinal markers and clinical endpoints. The model is fitted in the Maximum Likelihood framework by using an exact likelihood computation. The (possibly multiple) clinical endpoints are either repeated endpoints collected at repeated times or a time to event discretized into intervals of time. 
}

\details{
\tabular{ll}{
Package: \tab multLPM \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-06-12\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
  Viviane Philipps, Cecile Proust-Lima
}

%\references{
%}
