% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredEC.R
\name{PredEC}
\alias{PredEC}
\title{Probability of continuous event in the prediction interval ]s,s+t] for a specific subject}
\usage{
PredEC(num, s, t, model, Y, D, data, ID, var.time, RE = "block-diag",
  BM = "diag", breaks, B)
}
\arguments{
\item{num}{subject ID for whom we want to predict}

\item{s}{landmark s, lower bound of the prediction interval}

\item{t}{horizon t defining the upper bound of the prediction interval}

\item{model}{a JointMult model}

\item{Y}{a list of \code{multlcmm} objects. Each multlcmm object defines the outcomes and longitudinal structure of one dimension of the model. The number of outcomes, covariates or random effects can differ between dimensions.}

\item{D}{a list of two-sided formula defining the event part of the model. The left side should be survBreak(T0,T,Event) with T0 the entry time, T the event time and Event the event indicator}

\item{data}{data.frame containing the observations and variables}

\item{ID}{name of the subjects'ID variable}

\item{var.time}{a character vector indicating the name of the time variable of each dimension. The scales of these different time variables should be the same.}

\item{RE}{an indicator of the random effect structure between dimensions. Should be either "block-diag" for independent random effects between dimensions (the internal structure being defined in the multlcmm object) or "full" for correlated random effects between dimensions. Default is "block-diag".}

\item{BM}{in the case where Brownian motions are included in the multlcmm objects, an indicator of the correlation of the Brownian motions. Should be "diag" for independence, or "full" for correlated Brownian motions between dimensions. Default is "diag".}

\item{breaks}{a vector specifying the break points for event time's discretization}

\item{B}{optional vector containing values for the parameters}
}
\value{
the probability of continuous event in the prediction interval ]s,s+t] for a specific subject
}
\description{
Probability of continuous event in the prediction interval ]s,s+t] for a specific subject
}
\author{
Cecile Proust-Lima, Viviane Philipps and Tiphaine Saulnier
}
