% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uacv.R
\name{uacv}
\alias{uacv}
\title{Estimation of the Universal Approximate Cross Validation (UACV) criterion for a joint model}
\usage{
uacv(model, Y, D, data, var.time, RE = "block-diag", BM = "diag", B, posfix,
  breaks = NULL, delayed = TRUE)
}
\arguments{
\item{model}{a JointMult model}

\item{Y}{a list of \code{multlcmm} objects}

\item{D}{a list of two-sided formula defining the event part of the model}

\item{data}{data.frame containing the observations and variables}

\item{var.time}{a character vector indicating the name of the time variables}

\item{RE}{an indicator of the random effect structure between dimensions}

\item{BM}{an indicator of the correlation of the Brownian motions}

\item{B}{vector cntaining initial values for the parameters}

\item{posfix}{optional vector specifying the indices in vector B of the parameters that are not estimated}

\item{breaks}{optional vector specifying the break points in the case where the event time is discretized}

\item{delayed}{logical vector indicating if delayed entry should be accounted for}
}
\value{
the value of the Universal Approximate Cross Validation criterion
}
\description{
Estimation of the Universal Approximate Cross Validation (UACV) criterion for a joint model
}
\references{
Commenges D, Proust-Lima C, Samieri C, Liquet B.
A universal approximate cross-validation criterion for regular risk functions.
The International Journal of Biostatistics 2015; 11(1): 51–67.
doi: 10.1515/ijb-2015-0004
}
\author{
Cecile Proust-Lima and Viviane Philipps
}
